pyproject.toml
[tool.poetry]
name = "streamlit-multipage-template"
version = "0.1.0"
description = ""
authors = ["Krzysztof Olszewski <hello@retryjoin.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
geemap = "^0.32.1"
geopandas = "^0.14.4"
jupyter-server-proxy = "^4.2.0"
leafmap = "^0.34.3"
nbserverproxy = "^0.8.8"
owslib = "^0.31.0"
streamlit = "^1.36.0"
altair = "5.2.0"
attrs = "23.2.0"
blinker = "1.7.0"
cachetools = "5.3.2"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
gitdb = "4.0.11"
gitpython = "3.1.41"
google-analytics-data = "0.18.2"
google-api-core = "2.15.0"
google-api-python-client = "2.115.0"
google-auth = "2.26.2"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.0"
googleapis-common-protos = "1.62.0"
grpcio = "1.60.0"
grpcio-status = "1.60.0"
httplib2 = "0.22.0"
idna = "3.6"
importlib-metadata = "7.0.1"
jinja2 = "3.1.3"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
markdown-it-py = "3.0.0"
markupsafe = "2.1.4"
matplotlib = "3.9"
mdurl = "0.1.2"
numpy = "1.26.3"
oauthlib = "3.2.2"
packaging = "23.2"
pandas = "2.2.0"
pillow = "10.2.0"
proto-plus = "1.23.0"
protobuf = "4.25.2"
pyarrow = "14.0.2"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pydeck = "0.8.1b0"
pygments = "2.17.2"
pyparsing = "3.1.1"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
referencing = "0.32.1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rich = "13.7.0"
rpds-py = "0.17.1"
rsa = "4.9"
searchconsole = "0.0.4"
six = "1.16.0"
smmap = "5.0.1"
streamlit-elements = "0.1.0"
tenacity = "8.2.3"
toml = "0.10.2"
toolz = "0.12.0"
tornado = "6.4"
typing-extensions = "4.9.0"
tzdata = "2023.4"
tzlocal = "5.2"
uritemplate = "4.1.1"
urllib3 = "2.1.0"
validators = "0.22.0"
watchdog = "3.0.0"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
